+++ provider/pom.xml
@@ -47,12 +47,7 @@
 			<artifactId>lombok</artifactId>
 			<scope>provided</scope>
 		</dependency>
-		
-		<dependency>
-			<groupId>org.testng</groupId>
-			<artifactId>testng</artifactId>
-			<version>7.4.0</version>
-		</dependency>
+
 	</dependencies>
 
 	<build>
