+++ provider/src/main/java/com/afs/ezra/simpletype/provider/leaderboard/model/TypingTest.java
@@ -5,12 +5,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @NoArgsConstructor
 @Getter
 @Setter
-@ToString
 public class TypingTest {
 	@JsonProperty("text[]")
 	public String text;
