+++ provider/src/main/java/com/afs/ezra/simpletype/provider/leaderboard/service/LeaderboardService.java
@@ -3,13 +3,14 @@ package com.afs.ezra.simpletype.provider.leaderboard.service;
 import java.io.IOException;
 import java.util.List;
 
-import com.afs.ezra.simpletype.provider.common.HttpException;
 import com.afs.ezra.simpletype.provider.leaderboard.model.LeaderboardPlaceView;
 import com.afs.ezra.simpletype.provider.leaderboard.model.TypingTest;
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.databind.JsonMappingException;
 
 public interface LeaderboardService {
 	
-	LeaderboardPlaceView postLeaderboardScore(TypingTest testData) throws  IOException, HttpException;
+	LeaderboardPlaceView postLeaderboardScore(TypingTest testData) throws JsonParseException, JsonMappingException, IOException;
 
-	List<LeaderboardPlaceView> getTopLeaderboard(int length) throws HttpException;
+	List<LeaderboardPlaceView> getTopLeaderboard(int length);
 }
