+++ provider/src/main/java/com/afs/ezra/simpletype/provider/themes/model/ThemeImpl.java
@@ -11,7 +11,6 @@ import javax.persistence.Table;
 
 import org.springframework.beans.BeanUtils;
 
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -21,7 +20,6 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-@EqualsAndHashCode
 public class ThemeImpl implements Theme{
 	
 	public ThemeImpl(Theme theme) {
