+++ provider/src/main/java/com/afs/ezra/simpletype/provider/themes/repo/ThemeRepository.java
@@ -1,7 +1,6 @@
 package com.afs.ezra.simpletype.provider.themes.repo;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -9,7 +8,7 @@ import com.afs.ezra.simpletype.provider.themes.model.ThemeImpl;
 
 public interface ThemeRepository extends JpaRepository<ThemeImpl, Long> {
 	
-	public Optional<ThemeImpl> findThemeByName(String name);
+	public ThemeImpl findThemeByName(String name);
 	
 	public List<ThemeImpl> findAll();
