+++ provider/src/test/java/com/afs/ezra/simpletype/provider/ProviderApplicationTests.java
@@ -1,0 +1,13 @@
+package com.afs.ezra.simpletype.provider;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProviderApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
