--- webapp/src/main/resources/application-test.properties
@@ -1,2 +1,0 @@
-server.port=8080
-test.baseUrl=http://localhost:8080
